RT 1237 - Table-in-table bug
The problem is that when tables are nested, at least some of the content
in the containing table is truncated. This commit adds a table stack.
When a new table is found, the context for the previous one (if any) is
pushed onto the stack. When the end tag is found, the context is
restored from the stack.