Handle <th> like <td>
That is, create a {headers} key in the <tr> hash. This contains the <th>
hashes for the row, or undef if the position contains a <td>. Similarly,
the {cells} key contains an undef for a <th>. Trailing undefs in both
are removed, as is the whole tag if it is empty.

A true value for the {classic} argument to new() restores the old
behavior. The {headers} key in the <table> hash gets populated either
way.